+++ sniffer/tui/display.py
@@ -5,11 +5,14 @@ from time import sleep
 from datetime import datetime
 from rich.text import Text
 from rich.live import Live
+from rich.tree import Tree
 from rich.align import Align
 from rich.panel import Panel
 from rich.layout import Layout
+from rich.columns import Columns
 
 from ..utils import constants
+from ..parser.textutils import console
 from ..__version__ import __version__
 from ..network.engine import SnifferEngine
 from ..utils.funcutils import get_commit_hash
@@ -53,7 +56,7 @@ class Footer:
                f'Version: v{__version__}'
 
         return Panel(Text(text, justify="center"),
-                     title="[red]Information[/red]",
+                     title="[red]Status[/red]",
                      title_align="center",
                      border_style='bold red',
                      style="cyan")
@@ -64,7 +67,7 @@ class Footer:
                f'seconds...'
 
         return Panel(Align.center(text),
-                     title="[red]Information[/red]",
+                     title="[red]Status[/red]",
                      title_align="center",
                      border_style='bold red',
                      style="cyan")
@@ -94,47 +97,56 @@ class SidePanel:
 
 
 class Body:
-    def __init__(self, analyzer):
+    def __init__(self, analyzer: list[PacketAnalyzer], max_height: int):
         self.packets = analyzer
+        self.max_height = max_height
 
     def __rich__(self):
-        text_list = []
+        packet_list = []
+
         for analyzer in self.packets:
             formatted_packet = [
                 f'▷ [b magenta]#{analyzer.packet_count}[/b magenta]\t'
-                f'[green]{analyzer.source_ip}[/green] ⟶ '
-                f'[red]{analyzer.dest_ip}[/red] | '
-                f'[green]HTTP Version:[/green]\t'
+                f'[b green]{analyzer.source_ip}[/b green] ⟶ '
+                f'[b red]{analyzer.dest_ip}[/b red] | '
+                f'[b green]HTTP Version:[/b green]\t'
                 f'[b]{analyzer.http_version}[/b]'
             ]
+
             # This means it's an HTTP response
             if analyzer.http_verb is None:
-                formatted_packet.append(f' | [green]Status code:[/green]\t'
+                formatted_packet.append(f' | [b green]Status code:[/b green]\t'
                                         f'[b]{analyzer.status_code}[/b]\t'
                                         f'[b purple][Response][/b purple]')
             else:
-                formatted_packet.append(f' | [green]Method:[/green]\t'
+                formatted_packet.append(f' | [b green]Method:[/b green]\t'
                                         f'[b]{analyzer.http_verb}[/b]\t'
-                                        f'[b magenta][Request][/b magenta]'
-                                        f'\n\t[b magenta]Path:\t[/b magenta]'
-                                        f'{analyzer.request_path}')
+                                        f'[b magenta][Request][/b magenta]')
 
-            for header in analyzer.http_headers:
-                if not header:
-                    break
+            tree = Tree(''.join(formatted_packet),
+                        guide_style='underline2 cyan')
 
-                try:
-                    text = header.decode('utf-8')
-                except UnicodeDecodeError:
-                    text = header
+            if analyzer.http_verb is not None:
+                path = tree.add('⚫[b magenta]Path[/b magenta]')
+                path.add(f'{analyzer.request_path}')
 
-                formatted_packet.append(
-                    f'[b purple]\n\tHeaders:[/b purple] {text}'
-                )
-            formatted_packet.append(f'\n\tContent: {analyzer.http_body}\n')
-            text_list.append(''.join(formatted_packet))
+            headers = tree.add('⚫[b magenta]Headers[/b magenta]')
+            for key, value in analyzer.http_headers:
+                headers.add(f'[b green]{key}:[/b green]{value}')
 
-        return Align.left(''.join(text_list))
+            body = tree.add('⚫[b magenta]Body[/b magenta]')
+
+            if not analyzer.http_body:
+                body.add('[b purple]<empty>[/b purple]')
+            else:
+                body.add(f'{analyzer.http_body}')
+
+            packet_list.append(Panel(tree,
+                                     expand=True,
+                                     border_style='dim white',
+                                     height=analyzer.packet_height))
+
+        return Columns(packet_list, expand=True)
 
 
 class Banner:
@@ -243,7 +255,8 @@ def make_layouts(args: argparse.Namespace, sniffer: SnifferEngine):
 
 async def update_and_refresh(layout: Layout,
                              live: Live,
-                             packets: list[PacketAnalyzer]) -> None:
+                             packets: list[PacketAnalyzer],
+                             max_height: int) -> None:
     """Update the screen with the new information and refresh it.
 
     Parameters
@@ -254,9 +267,27 @@ async def update_and_refresh(layout: Layout,
         The context manager object which orchestrates the visible areas
     packets : list[PacketAnalyzer]
         A list containing parsed information from sniffed packets
+    max_height : int
+        The maximum height to be used by the layout
     """
-    layout.update(Body(packets))
+    layout.update(Body(packets, max_height))
     live.refresh()
+
+
+async def fit_packets(packets: list[PacketAnalyzer], max_height: int) -> None:
+    """Fit the packets to the screen by height.
+
+    Parameters
+    ----------
+    packets : list[PacketAnalyzer]
+        The list of HTTP packets
+    max_height : int
+        The maximum height allowed to be used for displaying the packets on
+        the screen
+    """
+    cumulative_sum = sum((packet.packet_height for packet in packets))
+    while len(packets) > 1 and cumulative_sum > max_height:
+        packets.pop(0)
 
 
 async def render(args: argparse.Namespace, sniffer: SnifferEngine):
@@ -266,22 +297,29 @@ async def render(args: argparse.Namespace, sniffer: SnifferEngine):
     with Live(banner,
               screen=True,
               redirect_stderr=False,
-              auto_refresh=False) as live:
+              auto_refresh=False,
+              transient=True,
+              console=console) as live:
         # The coolest thing you will see today
         intro(banner, live)
 
         live.update(panel, refresh=True)
         try:
             packets = []
+
             async for analyzer in sniffer.sniff(args.count):
-                if len(packets) == 8:
-                    packets = [analyzer]
-                else:
-                    packets.append(analyzer)
+                # 3 is footer + 1 is header + 2 for the borders (top & bottom)
+                max_height = console.height - 6
+
+                packets.append(analyzer)
+                await fit_packets(packets, max_height)
 
                 await asyncio.gather(*(
                     asyncio.create_task(
-                        update_and_refresh(layout['body'], live, packets)),
+                        update_and_refresh(layout['body'],
+                                           live,
+                                           packets,
+                                           max_height)),
                     asyncio.create_task(
                         write_output_to_file(sniffer.file_handle, analyzer))
                 ))